WIP: Parser for verbose parsing.
This is just the parsing part, and understands a Base Tx with SECP256K1
inputs and outputs; places where we should be tallying up or otherwise
keeping values for the prompts have debug PRINTFs, and the function to
actually hash data into a transaction hash is not filled in. The
corresponding test in tests uses a payload taken directly from the
serialization docs, and the parser does pass on this input.

We also need much "we don't know how to display this, reject" calls
added.