Just show all StakeableLockOut outputs.
This is enough for a user to securely sign a transaction with locked
output UTXOs, albeit by punting the verification of lock times to the
user.

it would be better user experience for the ledger to validate a
constraint like "for each unlock time, there is less than or equal to
the funds with that lock time in the output as there is in the input",
possibly with a prompt to override if that constraint is violated. This
does require significantly more involved work on the ledger and may not
be possible in constant memory, so the fallback position here is
expedient and reasonable.