Correct the StakeableLock implementation.
As it turns out, while the documentation for the P-chain format suggests
that StakeableLockOut contains a TransferableOutput, it merely contains
an Output, which is presumedly always a SECP256K1TransferOutput, and the
situation is analogous for Inputs. This adds a transaction generated by
the avalanchejs library to the test suite, corrects the parser, and
corrects the hand-written test case.