Fix another test and lower the batch size
`sendCommandAndAccept` now will return anything the function argument
returns. It has a type parameter so any sort of thing can be returned.

Use this to return the hash and signatures from
`Avalanch.signTransaction`, and check both of them.

Lower the batch size because of new overflow issues. Evidently the app
is consuming more stack space with the new SDK.